Move class to upper level
