+++ vck-openid/src/commonMain/kotlin/at/asitplus/wallet/lib/openid/JwsHeaderClientIdScheme.kt
@@ -1,0 +1,21 @@
+package at.asitplus.wallet.lib.openid
+
+import at.asitplus.signum.indispensable.josef.JwsHeader
+import at.asitplus.wallet.lib.agent.KeyMaterial
+import at.asitplus.wallet.lib.jws.JwsHeaderIdentifierFun
+
+class JwsHeaderClientIdScheme(val clientIdScheme: ClientIdScheme) : JwsHeaderIdentifierFun {
+    override suspend operator fun invoke(
+        it: JwsHeader,
+        keyMaterial: KeyMaterial,
+    ) = when (clientIdScheme) {
+        is ClientIdScheme.CertificateHash -> it.copy(certificateChain = clientIdScheme.chain)
+        is ClientIdScheme.CertificateSanDns -> it.copy(certificateChain = clientIdScheme.chain)
+        is ClientIdScheme.VerifierAttestation -> it.copy(
+            jsonWebKey = keyMaterial.jsonWebKey,
+            attestationJwt = clientIdScheme.attestationJwt.jws
+        )
+
+        else -> it.copy(jsonWebKey = keyMaterial.jsonWebKey)
+    }
+}
